Scanners moved to pkg root
