+++ pkg/scanner/aws_ec2.go
