+++ pkg/scanner/aws_ec2_test.go
