+++ pkg/scanner/aws_ecr.go
