+++ pkg/scanner/aws_ecr_test.go
