+++ pkg/scanner/aws_s3.go
