+++ pkg/scanner/aws_s3_test.go
