+++ pkg/scanner/fs_scanner.go
