+++ pkg/scanner/fs_scanner_test.go
