+++ pkg/scanner/scanner.go
