+++ pkg/scanner/scanner_mock.go
