+++ pkg/scanner/scanner_test.go
