+++ pkg/source/source.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/shaharia-lab/teredix/pkg"
 	"github.com/shaharia-lab/teredix/pkg/config"
-	"github.com/shaharia-lab/teredix/pkg/source/scanner"
+	"github.com/shaharia-lab/teredix/pkg/scanner"
 
 	"github.com/aws/aws-sdk-go-v2/service/resourcegroupstaggingapi"
